src/requirements.txt
pandas
numpy
requests
matplotlib
seaborn
kaggle
kagglehub
python-dotenv
